requirements.txt
deep_translator==1.11.4
fitz==0.0.1.dev2
Flask==2.2.2
opencv_python==4.8.0.76
pytesseract==0.3.10
python_docx==0.8.11
python_pptx==0.6.22